Add getLeadsByActionTaken query
In order for users to be able to filter complaints by action taken in NATCom,
a function to return the list of lead_identifiers that belong to case_files
that have had that action taken on them was needed. Current implementation
aims to avoid raw querying the database via TypeORM, but the result is 3
separate calls. Refactoring to a single query handled with joins may be a
good idea, however the impact of raw queries can cause problem with some
ORMs. Worth looking into, for now this follows existing pattern in the code
base.